refactor: CLDSRV-30 rename locationKeysSanityCheck to locationKeysHaveChanged
The function name now reflects what it is returning (a boolean) which
is clearer in the new context where it will be used. The jsdoc was
also improved and fixed.
