+++ lib/api/completeMultipartUpload.js
@@ -20,8 +20,8 @@ const { metadataValidateBucketAndObj } = require('../metadata/metadataUtils');
 const { skipMpuPartProcessing } = require('../data/external/utils');
 const locationConstraintCheck
     = require('./apiUtils/object/locationConstraintCheck');
-const locationKeysSanityCheck
-    = require('./apiUtils/object/locationKeysSanityCheck');
+const locationKeysHaveChanged
+    = require('./apiUtils/object/locationKeysHaveChanged');
 
 const logger = require('../utilities/logger');
 
@@ -362,9 +362,7 @@ function completeMultipartUpload(authInfo, request, log, callback) {
                 //   metadata keys, which are likely to have failed in
                 //   the previous MPU completion attempt
                 //
-                const onlyDifferentLocationKeys = locationKeysSanityCheck(
-                    objMD.location, dataLocations);
-                if (!onlyDifferentLocationKeys) {
+                if (!locationKeysHaveChanged(objMD.location, dataLocations)) {
                     log.info('MPU complete request replay detected', {
                         method: 'completeMultipartUpload.storeAsNewObj',
                         bucketName: destinationBucket.getName(),
